+++ DodgerGame/src/main/java/Entities/Projectiles/SmallAsteroid.java
@@ -4,14 +4,14 @@ package Entities.Projectiles;
  * @Author Olle Westerlund
  */
 public class SmallAsteroid extends Projectile {
-    private double damage; //The number of damage this asteroid does.
+    private int damage; //The number of damage this asteroid does.
 
     public SmallAsteroid() {
         super(5);
         this.damage = 20;
     }
 
-    public double getDamage() {
+    public int getDamage() {
         return damage;
     }
 }
